feat(vote): add revision column
* add revision column to Vote and Opinion
* refactor DataFlag: create DataSubset class also used by Opinions
