--- chimedb/dataflag/opinion/__init__.py
@@ -1,8 +1,0 @@
-from .orm import (
-    DataFlagClient,
-    DataFlagOpinionType,
-    DataFlagOpinion,
-    DataFlagVote,
-    MediaWikiUser,
-)
-from .vote import VotingJudge
