+++ chimedb/dataflag/orm.py
@@ -1,5 +1,5 @@
 """
-Table definitions for the CHIME data flagging tools
+Table definitions for the CHIME data flagging (and opinion) tools
 """
 # === Start Python 2/3 compatibility
 from __future__ import absolute_import, division, print_function, unicode_literals
@@ -8,10 +8,12 @@ from future.builtins.disabled import *  # noqa  pylint: disable=W0401, W0614
 
 # === End Python 2/3 compatibility
 
-from chimedb.core.orm import base_model, JSONDictField
+from .base_models import DataSubset, DataSubsetType
+
+from chimedb.core.orm import base_model, JSONDictField, EnumField
+from chimedb.core.exceptions import ValidationError
 from chimedb.mediawiki import MediaWikiUser
 
-import numpy as np
 import peewee as pw
 
 # Logging
@@ -27,7 +29,53 @@ _logger.addHandler(logging.NullHandler())
 # ====================================
 
 
-class DataFlagType(base_model):
+class DataRevision(base_model):
+    """
+    Revision of offline pipeline that produced the data.
+    """
+
+    name = pw.CharField(max_length=32, unique=True)
+    description = pw.TextField(null=True)
+
+
+class DataFlagClient(base_model):
+    """Client used to create a data flag.
+
+    Attributes
+    ----------
+    client_name : str
+        Name of the client software.
+    client_version : str
+        Version string.
+    """
+
+    client_name = pw.CharField()
+    client_version = pw.CharField()
+
+
+class DataFlagVote(base_model):
+    """
+    A Vote that resulted in the translation of opinions to flags.
+
+    Attributes
+    ----------
+    time : float
+        Unix time at vote creation.
+    mode : str
+        Voting mode name. They are implemented in `chimed.dataflag.opinion.vote`.
+    client : DataFlagClient
+        Client used to vote.
+    """
+
+    max_len_mode_name = 32
+
+    time = pw.FloatField()
+    mode = pw.CharField(max_length=max_len_mode_name)
+    client = pw.ForeignKeyField(DataFlagClient)
+    revision = pw.ForeignKeyField(DataRevision, backref="votes")
+
+
+class DataFlagType(DataSubsetType):
     """The type of flag that we are using.
 
     Attributes
@@ -40,12 +88,10 @@ class DataFlagType(base_model):
         An optional JSON object describing how this flag type is being generated.
     """
 
-    name = pw.CharField(max_length=64, unique=True)
-    description = pw.TextField(null=True)
-    metadata = JSONDictField(null=True)
+    pass
 
 
-class DataFlag(base_model):
+class DataFlag(DataSubset):
     """A flagged range of data.
 
     Attributes
@@ -78,67 +124,7 @@ class DataFlag(base_model):
     """
 
     type = pw.ForeignKeyField(DataFlagType, backref="flags")
-
-    start_time = pw.DoubleField()
-    finish_time = pw.DoubleField(null=True)
-
-    metadata = JSONDictField(null=True)
-
-    @property
-    def instrument(self):
-        """The instrument the flag applies to."""
-        if self.metadata is not None:
-            return self.metadata.get("instrument", None)
-        else:
-            return None
-
-    @property
-    def freq(self):
-        """The list of inputs the flag applies to. `None` if not set.
-        """
-        if self.metadata is not None:
-            return self.metadata.get("freq", None)
-        else:
-            return None
-
-    @property
-    def freq_mask(self):
-        """An array for the frequencies flagged (`True` if the flag applies).
-        """
-
-        # TODO: hard coded for CHIME
-        mask = np.ones(1024, dtype=np.bool)
-
-        if self.freq is not None:
-            mask[self.freq] = False
-            mask = ~mask
-
-        return mask
-
-    @property
-    def inputs(self):
-        """The list of inputs the flag applies to. `None` if not set.
-        """
-        if self.metadata is not None:
-            return self.metadata.get("inputs", None)
-        else:
-            return None
-
-    @property
-    def input_mask(self):
-        """An array for the inputs flagged (`True` if the flag applies).
-        """
-        if self.instrument is None:
-            return None
-
-        inp_dict = {"chime": 2048, "pathfinder": 256}
-        mask = np.ones(inp_dict[self.instrument], dtype=np.bool)
-
-        if self.inputs is not None:
-            mask[self.inputs] = False
-            mask = ~mask
-
-        return mask
+    vote = pw.ForeignKeyField(DataFlagVote, backref="flags", null=True)
 
     @classmethod
     def create_flag(
@@ -204,3 +190,184 @@ class DataFlag(base_model):
             metadata=table_metadata,
         )
         return flag
+
+
+# Tables pertaining to the data flag opinions.
+# ============================================
+
+
+class DataFlagOpinionType(DataSubsetType):
+    """The type of opinion that we are using.
+
+    Attributes
+    ----------
+    name : string
+        Name of the type of opinion.
+    description : string
+        A long description of the opinion type.
+    metadata : dict
+        An optional JSON object describing how this opinion type is being generated.
+    """
+
+    pass
+
+
+class DataFlagOpinion(DataSubset):
+    """A persons opinion on flagging a range of data.
+
+    Attributes
+    ----------
+    type : DataFlagOpinionType
+        The type of flag.
+    user : MediaWikiUser
+        The user who made this opinion.
+    decision : str
+        Decision about the data: "good", "bad" or "unsure".
+    start_time, finish_time : double
+        The start and end times as UNIX times.
+    metadata : dict
+        A JSON object with extended metadata. See below for guidelines.
+    creation_time, last_edit : double
+        The time of creation and last edit as UNIX times.
+    revision : str
+        Name of data revision this opinion based on.
+
+    Notes
+    -----
+    To ensure that the added metadata is easily parseable, it should adhere
+    to a rough schema. The following common fields may be present:
+
+    `instrument` : optional
+        The name of the instrument that the flag opinion applies to. If not set,
+        assumed to apply to all instruments.
+    `freq` : optional
+        A list of integer frequency IDs that the flag opinion applies to. If not
+        present the flagging opinion is assumed to apply to *all* frequencies.
+    `inputs` : optional
+        A list of integer feed IDs (in cylinder order) that the flagging opinion applies
+        to. If not present the flagging opinion is assumed to apply to *all* inputs. For
+        this to make sense an `instrument` field is also required.
+
+    Any other useful metadata can be put straight into the metadata field,
+    though it must be accessed directly.
+    """
+
+    type = pw.ForeignKeyField(DataFlagOpinionType, backref="opinions")
+    user = pw.ForeignKeyField(MediaWikiUser, backref="opinions")
+    decision = EnumField(["good", "bad", "unsure"])
+
+    creation_time = pw.DoubleField()
+    last_edit = pw.DoubleField()
+    client = pw.ForeignKeyField(DataFlagClient)
+    revision = pw.ForeignKeyField(DataRevision, backref="opinions")
+
+    @classmethod
+    def create_opinion(
+        cls,
+        username,
+        creation_time,
+        decision,
+        opiniontype,
+        client_name,
+        client_version,
+        start_time,
+        finish_time,
+        revision,
+        freq=None,
+        instrument="chime",
+        inputs=None,
+        metadata=None,
+    ):
+        """Create a flagging opinion entry.
+
+        Parameters
+        ----------
+        decision : str
+            "good", "bad" or "unsure".
+        opiniontype : str
+            Name of opinion type. Must already exist in database.
+        start_time, finish_time : float
+            Start and end of flagged time.
+        username : str
+            Name of the user entering the opinion.
+        creation_time : double
+            Unix time when the opinion was entered. Default: current time.
+        freq : list, optional
+            List of affected frequencies.
+        instrument : str, optional
+            Affected instrument.
+        inputs : list, optional
+            List of affected inputs.
+        metadata : dict
+            Extra metadata to go with the opinion entry.
+        revision : str
+            Name of data revision this opinion based on.
+
+        Returns
+        -------
+        opinion : DataFlagOpinion
+            The opinion instance.
+        """
+
+        if not isinstance(decision, str) or decision not in cls.decision.enum_list:
+            raise ValidationError(
+                "Invalid value '%s' for 'decision'. Choose one of %s"
+                % (decision, cls.decision.enum_list)
+            )
+
+        table_metadata = {}
+
+        if freq is not None:
+            if not isinstance(freq, list):
+                raise ValueError("freq argument (%s) must be list.", freq)
+            table_metadata["freq"] = freq
+
+        if instrument is not None:
+            table_metadata["instrument"] = instrument
+
+        if inputs is not None:
+            if not isinstance(inputs, list):
+                raise ValueError("inputs argument (%s) must be list.", inputs)
+            table_metadata["inputs"] = inputs
+
+        if metadata is not None:
+            if not isinstance(metadata, dict):
+                raise ValueError("metadata argument (%s) must be dict.", metadata)
+            table_metadata.update(metadata)
+
+        # Get the flag
+        type_ = DataFlagOpinionType.get(name=opiniontype)
+
+        # Get the revision
+        revision = DataRevision.get(name=revision)
+
+        # Get the user
+        username = username[0].upper() + username[1:]
+        user_ = MediaWikiUser.get(user_name=username)
+
+        # Get the client
+        client, _ = DataFlagClient.get_or_create(
+            client_name=client_name, client_version=client_version
+        )
+
+        opinion = cls.create(
+            user=user_,
+            creation_time=creation_time,
+            last_edit=creation_time,
+            decision=decision,
+            type=type_,
+            client=client,
+            start_time=start_time,
+            finish_time=finish_time,
+            metadata=table_metadata,
+            revision=revision,
+        )
+
+        return opinion
+
+
+class DataFlagVoteOpinion(base_model):
+    """Many-to-many relationship between votes and opinions."""
+
+    vote = pw.ForeignKeyField(DataFlagVote)
+    opinion = pw.ForeignKeyField(DataFlagOpinion)
